requirements.txt
plyer
pyautogui
pystray
pillow
cryptography